+++ src/void/engine/frame.go
@@ -7,7 +7,6 @@ type Frame struct {
 	// time since the last frame was _requested_ in milliseconds.
 	DeltaMs float64
 	// time in UTC milliseconds.
-	NowMs float64
-	// to-do: move this here.
-	CanvasW, CanvasH uint32
+	NowMs            float64
+	CanvasW, CanvasH uint16
 }
